03_price_monitoring/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-pubsub==2.*
cloudevents==1.*
